requirements.txt
# Interactive Brokers Integration
ib_insync==0.9.86

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Configuration
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
